requirements.txt
aiotdlib==0.22.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
certifi==2023.11.17
Django==5.0.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
pydantic==1.10.14
python-dotenv==1.0.1
python-telegram-bot==20.7
redis==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
typing_extensions==4.9.0
ujson==5.9.0
